rename git event
